requirements.txt
Pillow==7.2.0
requests==2.24.0  